Make clash-cores build on ghc 8.8.4
The GHC 8.4.4 pipeline in the CI was failing due to not being
able to build clash-cores. This has been hopefully fixed by relaxing
the version bounds for base, and only enabling -XNoStarIsType when
GHC is at least 8.6.

The test.sh script for CI has also been changed to expect 5 versions
instead of 4 (this should have been changed when clash-cores was
added to the list of versions checked).