Split test groups out into separate functions
Each test group in the unit tests of clash-cores was very heavily
copy-pasted, while using the same basic function for tests with /
without a slave latch.

The separated test groups make things slightly more manageable,
but there is still room to remove a fair amount of duplication in
the test suite.