Move tests into one function again
The tests are once again definined in a single function, because
there was a lot of boilerplate from the fact only the testLatch
and testNoLatch functions differed between tests.

Currently, I think that changing the general test functions in the
module might help better refactor the test tree, so something might
be done there. Regardless, the test module is approximately 100
lines shorter, which is at least an improvement.