Extract common behaviour in test functions
The functions to simulate an SPI chip and sample the results were
implemented near-identically. By replacing them with a function
sampleSPI, and implementing the other functions in terms of this,
the test suite becomes much simpler to understand.

The problem still remains that sampleSPI is not general enough
to handle cases where there are multiple slaves, so the
testMasterMultiSlave function is still overly verbose.