Fix clash-cores tests (#908)
Another set of changes merged into master between the test refactoring
branch and this branch broke the unit tests. This commit fixes
everything in preparation for solving the issue at hand.

The documentation provided for SPI mode was also found to be
incorrect, as it conflated leading/trailing and rising/falling when
talking about edges. It has been rewritten to be more clear about what
the different modes mean.

Parts of the SPI module which were predicated on the mode have
been changed to use auxiliary functions which more clearly show
what is being tested (e.g. sampleOnRisingEdge tells whether the
mode samples on the rising edge of SCLK).

A lot of the cabal file was also reduplicated, so it has been changed
to a cabal 2.2 format to allow the common stanza to be used.