Use QuickCheck for testing SPI (#908)
The tests for SPI were currently very repetitive and needlessly rigid,
so they have been redefined to use QuickCheck. It is also now possible
to test SPI with a simple transfer (one word sent repeatedly) or with a
transfer that cycles through a list of values to send (useful for
testing multi-word messages).

Noticing similarities between the tests before this PR, and the tests
defined while working on this PR led to some tests being combined into
more general tests which show both. This makes the test suite a lot
more manageable, as most specific use cases are conceptually just
a combination of a few simpler tests.

The tests for multiple slaves are left unchanged, but moved to a new
module. Ideally later someone will think of a general test driver
function which allows for an arbitrary number of slaves to be used.