Generate enums in VHDL
The VHDL backend now generates enums for Sum's (the netlist type for
ADTs where each constructor has only zero-width fields) instead of
generating a bitvector for the type. This makes it somewhat easier to
follow the generated HDL, as it preserves more user-given names in the
final output.

Enums are generated by default. This can be disabled using the new
-fclash-no-render-enums flag, which exposes the old behaviour.