Added zeroWidthSpec transformation
A new transformation has been defined, which replaces lambda-bound variables
for a zero-width value with the value itself in the body of the lambda. e.g.

```haskell
data AB = A | B

ab :: KnownNat n => Index n -> AB -> AB
ab n A = if n >  0 then A else B
ab n B = if n == 0 then B else A
```

would see the values of n in the function body be replaced with 0 (the only
possible value of the type `Index 1`)

```haskell
ab _ A = if 0 > 0 then A else B
ab _ B = if 0 == 0 then B else A
```

the normalizer is then free to further normalize this to

```haskell
ab _ _ = B
```

where previously it would not be able to improve this during normalization.

Fixes #508