clash-testsuite: three fixes (#1930)
- The logic for `TestExitCode` in `runFailingProgram` was wrong:
  - `TestExitCode`: correct
  - `NoTestExitCode`: correct
  - `TestSpecificExitCode`:
     - A program succeeding will always report error, even if
     `TestSpecificExitCode 0` is used.
     - A program failing will always report success, even if the
     specific exit code did not match the intended one.

Because all current tests use `TestExitCode`, the wrong logic was not a
problem in practice so far.

- The error message for `runProgram` with its `stdF` argument set to
`True` would be the wrong error message for that failure condition, even
mentioning an incorrect return code of the program!

- We have a test to check that a program, or alternatively specificly
VVP, does not produce anything on stdout. However, this was in many
places named and documented as if it checked that it _does_ produce
something on stderr. Names and documentation adapted so they now
correspond to the actual functionality (which was correct).

Finally, some minor tuning to messages.