Selectively use tools with flags
It is now possible to selectively enable tests which use a
specific vendor tool by passing a flag to the testsuite driver,
e.g. --verilator for running tests with verilator. If a tool is not
given it's tests pass (as the tasty API does not allow deciding to
skip a test inside IsTest.run). If the tool is not found the error
is not tidied up instead of showing the contents of stderr.

This is a somewhat crude first attempt at addressing #2012.