Split overly sequential Driver code
The `generateHDL` function in the driver was written in a way which
made it difficult to concurrently compile entities. This needed
some rearranging to expose an (internal) API with less coupling on
things like tracking compilation time etc.

Some small refactoring is also performed, moving larger blocks /
bindings with a clear purpose to their own top-level definitions.