Define ClashEnv and ClashDesign
Common values which are passed around sections of the compiler are
now split into two types:

  * ClashEnv, which defines information which is more less specific
    to a design and used for later lookup (such as information about
    known primitives, or the command line options used to run clash)

  * ClashDesign, which defines information more relevant to the
    specific design being compiled (such as entity information and
    the bindings which make up the design)

The distinction between these is slightly hand-wavy. Some things,
like the BindingsMap are in ClashDesign even though it contains all
global bindings. Likewise, some more design-specific things may be
included in ClashEnv if they are used often, as ClashEnv is more
commonly passed around. This is intended as less of an absolute
classification and more of a convenient way to group data in the
compiler (while reducing the number of arguments given to functions).