`clash-testsuite`: make `hdlLoad` a list
This allows us finer control over which simulators to load HDL into,
just like we did with `hdlSim` in PR #2257.

The tests `shouldfail.Verification.NonTemporal{PSL,SVA}` accidentally
ran Vivado and the make stage of Verilator. This no longer happens.

This means we can remove the Vivado test filtering on `expectSimFail`
since that was just an artifact of that, accidentally running a
GHDL-only test in Vivado.