`clash-testsuite`: Harmonize simulators, clean up
Control execution of _all_ simulators, their load stage and their simulate
stage, through `hdlLoad` and `hdlSim`.

Verilator used to be configured through `verilate` (`SimAndVerilate`,
`VerilateOnly` and `SimOnly`), but is now configured like the rest.

At first, we had one tool for each HDL language: VHDL ran in GHDL,
Verilog ran in Icarus Verilog and SystemVerilog ran in ModelSim. Then we
added Verilator, which runs both Verilog and SystemVerilog, and this
exception to the structure was coded with the mentioned `verilate`
option. Now we also added Vivado, which runs VHDL and Verilog (and which
we'll make to run SystemVerilog as well later), and it is time for a
reorganization.

Unrelated to this, the function `Test.Tasty.Clash.hdlFiles` is removed
because it turns out it was unused since commit 5f9dd2682 of 2018-11-26,
where its use was replaced by a shell glob.