Use `infinite-list` to eliminate magic numbers
In primitive templates, we often refer to inputs by their index in the
list of all inputs. This is cumbersome and error-prone, and makes
maintenance a headache. GHC 9.2 is very quick on reporting an
`incomplete-uni-patterns` warning, but by using a provably infinite
list, we can graft an arbitrary number of binders from that list without
warnings and name our indices with them.

All occurences of these magic numbers that I could find have been
replaced by named indices, except for where this is impossible (notably
`clash-lib/prims` which contains many, many more).

The `TemplateFunction`s in `clash-lib:Clash.Primitives.Intel.ClockGen` were
technically incorrect. It didn't matter though: the `used` field is not actually
currently used in Clash, and if the `valid` field is too optimistic, the
render function (last field) will still throw an error message to the
user. This different error message provides more information in several
ways, but it crucially lacks the location in the user code that
triggered the error, which is a significant disadvantage.
The problems were these:
  - `used` for `altpll` was overly broad, listing all arguments. Collectively,
    the two `TemplateFunction`s did need all arguments, though.
  - `used` for `alteraPll` lists 21 inputs but the function has only 6.
    Most likely the up-to-16-tuple that is one of the constraints was
    accidentally thought to be expanded into separate constraints.
  - `valid` for `alteraPllTemplate` doesn't check `bbResults` for the correct
    shape.
  - The `valid`s for the `QsysTF`s don't check the parsed constraint
    arguments for validity.
  - Conversely, they do  check the `name` argument and the results for
    validity, but these are only used by `altpllTF` and `alteraPllTF`
    and are also already checked there.

The `used` fields were fixed. The `valid` functions have superfluous checks
removed, but no checks were added. The net result is that the same error
messages are emitted as before this patch.