Add the 'KEEP' HDL attribute to VIO Probe ports
When using a VIO, you do not want the names of the probes to get
optimized away. By adding the `KEEP` HDL attribute to the expressions
of the probe ports, the synthesis tool will try not to optimize the
`wire` (verilog) or `signal` (vhdl) away.

Additionally, the component declaration in vhdl did not have the same
types for each port as the component instantiation (port map). All ports
are now `BitVectors`, which translate to `std_logic_vector`s.