Add loading of HDL to VIO tests
Before adding the `KEEP` attribute to `vioProbe` ports, clash did
not generate valid VHDL since the types of vioProbe declaration
and instantiation did not match. This has been fixed, but to make
sure these bugs do not occur again the circuits in `shouldwork`
are now loaded in Vivado.

The circuits in `shouldwork` also did not produce valid VHDL, as
the words `in` and `out` are reserved. The names of ports have
been updated to produce valid VHDL.