Find the clash executable
Note that it appears that the Cabal store does not need to be first in
the env var `GHC_PACKAGE_PATHS` (it is automatically appended because
the env var ends in a `:`). The fact that the path was wrong for GHC 9.8
yet it worked fine clued us in on it being superfluous.