New design works in sim, not in hardware.
There are currently issues with the Probe path (how??)
Also, the record builder is way too complex. And processor as well due
to the bypass being able to give backpressure. New idea is to switch
bypass back to `CSignal` and have a separate FSM in the builder that
handles caching of the metadata.