Use ddrForwardClock
All arguments passing in a DDR primitive have also gained an `Enable`,
and the DDR output primitive gets its inputs as a tuple rather than as
two arguments. This matches the type of the DDR primitives in
`clash-prelude` (especially after https://github.com/clash-lang/clash-compiler/pull/2833
is merged)

`ddrForwardClock` is only in Clash master, we no longer support Clash 1.8 in
this core.