Fix bug in interpacket gap inserter
While using the interpacket gap inserter as an example I came across
an error that was not caught by any tests.

Currently, if the receiver asserts backpressure on the last fragment
of a packet, the component starts inserting the interpacket gap,
resulting in a gap inside of a packet. This is obviously not correct,
but the id test does not care about this. Fixed this by only changing
the state to Forward if the receiver acknowledges the last fragment.

I recommend adding a test which verifies that there are no gaps inside
of packets (only if the input contains no gaps).