Use `instWithXilinxWizard` for `fromU32With`
The old Tcl template function set the Xilinx wizard property
`CONFIG.Has_ACLKEN` based on analysis of the `Enable` input: if Clash
could determine the component was always enabled, the property would be
set to `false` causing the wizard to omit the clock enable line. The new
method with `inst` does not currently support this, and the choice was
made to always have the wizard emit a clock enable line, which is
functionally correct. Synthesis might even remove the clock enable,
leading to the same optimal result, but this was not investigated, so
the synthesis result might be suboptimal in that respect.