fixup! Add `instWithXilinxWizard`
Some of your `String` use could be replaced by `Text`, which chooses the better
type, but also splices quicker into a `Doc`, since a `Doc` is a `Text` and if
you splice `String` into it with `__di`, the very first thing that'll happen is
it'll invoke `Text.pack` on that `String` (this is from memory, don't quote me
on this).

This also incorporates two changes I mentioned in the PR review since they
change as well.