Add `DiffClock` support to `inst`
`DiffClock` is only supported as an input of `inst`. There are now 2 ways to
use `DiffClock` for `inst`: `DiffClockPort` gets a single symbol for the names
of the ports and adds the suffixes '_p' and '_n' for the positive and negative
phase respectively. `NamedDiffClockPorts` gets two symbols for the port names.