Rework Nix infrastructure (#46)
Restructures everything Nix related to be more similar to how
clash-protocols and clash-circuits have it setup.

This means it now uses flakes, removes all the 'old style' Nix files and
applies the GHC patches mentioned in clash-cores.cabal prior to this
commit.

All of this is very similar to how it is done in:
https://github.com/clash-lang/clash-protocols/pull/164