Add devshell/package caching (#48)
Compiles the Nix package and devshells (including HLS) for clash-cores and uploads them to the local binary cache and cachix.
These runners are hosted locally and Diepenheim and only activate on commits to the main branch. I have tested both pushing to the local binary cache and cachix.

This will cache the version of clash-compiler and clash-protocol as well, but only the ones pinned by the flake.lock. Those projects will need their own cache in-order to support different versions too.